backend/requirements.txt
fastapi
uvicorn
pydantic-settings
python-dotenv
python-multipart
pypdf
python-docx